Kind of working
After a lot of struggling, the code seems to be working now.  This is based on
comparison with the RADMC result.

The neighboring algorithm seems to be problematic, and needs to be corrected.
But that does not seems to help much.

The technical details in reemission (getting a new lambda) needs to be checked.

Also need to add suport for multiple dust species.

The input format needs to be refined.