Correction for dust-gas energy exchange
When the number of photons received by a cell is small, and if this cell is
expected to loose energy to the gas, previously the code cannot decide which
wavelength should this cell reemit, because the amount of energy to be emitted
it negative.

Now I only turn on the energy exchange if the energy to be reemitted is larger
than a certain amount.