Modified the dust-gas collisional heating
To get rid of the NaNs.

A variable grand_abundance added, which is the sume of the abundance of all the
species.  It is used in the dust-gas collisional heating/cooling rate
expression to accound for the fact that, when the dust temperature is extremely
low, everythign will stick to the dust and there will be no collision.

Also deleted two unused parameters.