A bug in calculating column density corrected
This happens for H2 and CO for which the initial abundance is very high, and
for cells for which the line connecting their center and the star cross the
cell just below it.  The column density of H2 and CO are overestimated.

In the corrected version the initial abundances of all the species of each cell
are set to zero.  The input initial abundances are only copied to the chemical
solver, not to each cell.  The cells only get their abundances after the
chemical solution.