OI and CII cooling changed back to before
On Jul 15, 2014 the critical column densities of them were increased from a few
times 1e20 to a few times 1e27.

I can' remember why, though the commit message at that time seems to indicate
that the change was not a solid one.

Now I change them back to a few times 1e20, based on the values in Hollenbach
1989.

The reason is that when the dust temperature of the midplane is high, O I can
stay in the gas phase, and can efficiently cool the gas if the critical column
density is high (hence essentially optically thin), so that the gas temperature
can be lower than the dust temperature, which may not be expected (though not
impossible; let's see what happens after the current change).

It is important to include the reason and context of a change in the commit
message (or the code comments).