Modify cdms.f90 to work with jpl format.
Energy level duplicates now checked with quantum number, rather than with energy values (which can be inaccurate).

Working on grid.f90 for better surface-edge-finding.

Modified ray_tracing.f90 to calculate partition function for the temperatures specified in the JPL documentation for easier comparison.  A bug was found (which does not affect results but do affect a log file).

In sub_trivials.f90 a binary_solve function is added.